refactor: HomeTopView 구현
HomeView의 상단 뷰 (타이틀, 버튼)을 subview로 구현
